+++ CHANGELOG.md
@@ -8,12 +8,13 @@ All notable changes to QETLAB will be documented in this file.
 - CopositivePolynomial: Creates a homogenous polynomial whose non-negativity is equivalent to copositivity of a given matrix.
 - EntangledSubspace: Constructs a basis of a bipartite r-entangled subspace of any dimension.
 - EntFormation: Computes the entanglement of formation of a 2-qubit state or a pure state.
+- IsCopositive: Determines whether or not a matrix is copositive.
 - MatsumotoFidelity: Computes the Matsumoto fidelity of two density matrices.
 - ParallelRepetition: Produces the coefficients of a parallel repetition of a nonlocal game, in full probability notation. Replaces the old NonlocalGameValue.m.
 - PolynomialAsMatrix: Creates a compact fully symmetric matrix representation of a polynomial.
 - PolynomialOptimize: Bounds the optimal value of a homogeneous polynomial on the unit sphere.
 - PolynomialSOS: Bounds the optimal value of a homogeneous polynomial on the unit sphere via the Sum-Of-Squares hierarchy.
-- RandomGraph: Generates a random graph.
+- RandomGraph: Generates the adjacency matrix of a random graph.
 - RandomPPTState: Generates a random density matrix with positive partial transpose, and optionally low rank.
 - helpers/asum_vector: Creates all vectors with binary entries adding to a given value. Used by AntisymmetricProjection.m.
 - helpers/asymind: Creates all vectors with strictly increasing permutations of an input vector.